+++ apps/web/src/common/career/CareerList.tsx
@@ -2,7 +2,7 @@ import { useGetUserCareerQuery } from '@/api/career/career';
 import { UserCareerType } from '@/api/career/careerSchema';
 import CareerItem from '@/common/career/CareerItem';
 import { DEFAULT_PAGE_INFO } from '@/common/career/constants';
-import MuiPagination from '@/domain/program/pagination/MuiPagination';
+import MuiPagination from '@/common/pagination/MuiPagination';
 import { convertCareerApiToUiFormat } from '@/utils/career';
 import { useState } from 'react';
